Code restructuring and task system implementation
The framework for task handling is now mostly in place
Thread support requires boost libraries for now, and will make compilation
on Windows significantly more challenging...

Various hacks put in place to ensure that rendering and event handling
always happen in the main thread.

The interface for accessing managers is now in place.
All public functions in the manager classes are required to be threadsafe,
while the protected and private functions may or may not be.
The manager interface now consists of all the public functions of the
manager class.

Restructuring and minor changes